update vault list
