+++ src/components/Logo/Logo.tsx
@@ -1,4 +1,4 @@
-import Image from 'next/image'
+import Image from 'next/future/image'
 import Link from 'next/link'
 import React from 'react'
