+++ src/pages/farms/components/Actions.tsx
@@ -21,7 +21,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useWeb3React } from '@web3-react/core'
 import { BigNumber, Contract, ethers } from 'ethers'
 import { formatUnits, parseUnits } from 'ethers/lib/utils'
-import Image from 'next/image'
+import Image from 'next/future/image'
 import Link from 'next/link'
 import { default as React, useCallback, useState } from 'react'
 import { FarmWithStakedValue } from './FarmList'
