+++ src/pages/vaults/components/Modals/Modals.tsx
@@ -10,7 +10,7 @@ import { useVaultPrices } from '@/hooks/vaults/usePrices'
 import { decimate, exponentiate, getDisplayBalance, sqrt } from '@/utils/numberFormat'
 import { BigNumber, FixedNumber } from 'ethers'
 import { formatUnits, parseUnits } from 'ethers/lib/utils'
-import Image from 'next/image'
+import Image from 'next/future/image'
 import React, { useCallback, useMemo, useState } from 'react'
 import VaultButton from '../VaultButton'
 import VaultStats from '../Stats'
